models/requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
Flask==3.1.0
flatbuffers==25.1.21
gast==0.6.0
google-pasta==0.2.0
grpcio==1.69.0
h5py==3.12.1
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.5
keras==3.8.0
lazy-loader==0.4
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
networkx==3.4.2
numpy==2.0.2
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pillow==11.1.0
pip==21.2.3
protobuf==5.29.3
Pygments==2.19.1
PyJWT==2.10.1
requests==2.32.3
rich==13.9.4
scikit-image==0.25.0
scipy==1.15.1
setuptools==57.4.0
six==1.17.0
sqlparse==0.5.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-intel==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
tifffile==2025.1.10
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
gradio>=4.0.0
# Add any other dependencies your app needs
